non functoning attempt at recording traffic ligjhts
I suspect that the IDs dont matchnot sure how to fix that.

+++ Assets/TrafficLightSupervisor.cs
@@ -8,7 +8,7 @@ public class TrafficLightSupervisor : MonoBehaviour
 {
     private TrafficLightController[] m_TrafficLightControllers;
 
-    public enum trafficLightStatus :byte
+    public enum trafficLightStatus :int
     {
         IDLE,
         RED,
